moved addnoteaction and fixed controller javadoc
